修改了对房源面积的数据存储长度

+++ zmcommon/src/main/java/com/zm/service/entity/House.java
@@ -65,7 +65,7 @@ public class House {
 	private Integer room;
 	
 	@Column(name = "area")
-	@ColumnType(jdbcType = JdbcType.TINYINT)
+	@ColumnType(jdbcType = JdbcType.INTEGER)
 	private Integer area;
 	
 	@Column(name = "toilet")
